django/requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.1
django-environ==0.9.0
django-leaflet==0.28.3
djangorestframework==3.13.1
idna==3.4
psycopg2==2.9.3
pycodestyle==2.9.1
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.12
